Moving all code to test Moving grid to resources
