--- src/main/java/pages/base/IPage.java
@@ -1,5 +1,0 @@
-package pages.base;
-
-interface IPage {
-
-}
