+++ src/test/java/org/selenide/examples/config/pages/base/IPage.java
@@ -1,0 +1,5 @@
+package org.selenide.examples.config.pages.base;
+
+interface IPage {
+
+}
