+++ src/test/java/org/selenide/examples/config/repositories/ESelectors.java
@@ -1,4 +1,4 @@
-package repositories;
+package org.selenide.examples.config.repositories;
 
 public enum ESelectors {
     SELECTOR_NAME_Q("q");
