+++ src/test/java/org/selenide/examples/tests/pages/GoogleSearchTest.java
@@ -1,22 +1,19 @@
-package org.selenide.examples.pages;
+package org.selenide.examples.tests.pages;
 
-import io.qameta.allure.Feature;
-import io.qameta.allure.Issues;
 import org.junit.Test;
-import pages.app.GoogleResultsPage;
-import pages.app.GoogleSearchPage;
+import org.selenide.examples.config.pages.app.GoogleResultsPage;
+import org.selenide.examples.config.pages.app.GoogleSearchPage;
 import ru.yandex.qatools.allure.annotations.Features;
 import ru.yandex.qatools.allure.annotations.Issue;
-import ru.yandex.qatools.allure.annotations.Step;
 import ru.yandex.qatools.allure.annotations.Stories;
-import tests.BaseTests;
+import org.selenide.examples.config.tests.BaseTests;
 
 import static com.codeborne.selenide.Selenide.open;
 
 public class GoogleSearchTest extends BaseTests {
     @Test
-    @Features("A")
-    @Stories({"A"})
+    @Features("Seraches for common phrases")
+    @Stories({"Google Basic Searches"})
     @Issue("JIRA-67")
     public void userCanSearchAnyKeywordAndResultsPageIsReturned() {
         GoogleSearchPage googleSearchPage = open(GoogleSearchPage.getAbsoluteUrl(), GoogleSearchPage.class);
