+++ src/test/java/org/selenide/examples/tests/suites/SubSuite.java
@@ -1,8 +1,8 @@
-package org.selenide.examples.suites;
+package org.selenide.examples.tests.suites;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
-import org.selenide.examples.pages.GoogleSearchTest2;
+import org.selenide.examples.tests.pages.GoogleSearchTest2;
 
 @RunWith(Suite.class)
