+++ src/test/resources/grid/grid/node1config.json
