+++ src/test/resources/grid/grid/startnode.bat
