+++ src/test/resources/grid/tip.txt
@@ -1,0 +1,1 @@
+This is local grid used for testing, do not use in production, instead use grid of docker or selenoid
